backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [tool.poetry.scripts]
# start = "uvicorn app.api.main:app --reload"